Implement the Repositories [Abstraction]...
